Add routing fixes from previous branch
Change bpgen bundle-size argument to float type

    Allows us to send bundles at rates less than 1 per second.
    Useful for testing.

Only initialize outducts in scheduler once

    Doing it twice destroys state: The linkIsUpTimeBased variable
    in the outductinfo structure is reset to false.

    This is used to track whether a link is logically up from the
    perspective of the contact plan (ignoring any physical link)
    changes from egress.

    Problem: If we get a new all-outducts-capabilities-telem
    message, then the outduct tracking data structures are
    reset, and this value is changed to be false.

    If the contact plan is not refreshed, then the scheduler will
    see this link as down per the contact plan. It will not update
    the router if the physical link goes up. This prevents the router
    from updating routes.

    We probably don't see this since we get contact plan updates every
    15 seconds in "production" testing.

    To fix this, only populate the maps on the first message and
    only later if the maps change.

Make routes based on physical link state

    Adds a data structure to track whether an outduct is up or not.
    When doing the routing, the router reads in the contact plan
    and makes the route based on the contact plan.

    Previously, the routing algorithm just used the contact plan, it
    had no concept of whether those links where physically up or down.

    Changing that behavior, here we prune (filter) the contact plan.
    For any current contact (current as in the contact is happening
    "now"), we delete it from the contact plan if the link is down.

    This contact plan is a local copy, so it's fine to modify it.

+++ common/bpcodec/apps/bpgen/include/BpGenAsyncRunner.h
@@ -32,7 +32,7 @@ public:
     BpGenAsyncRunner();
     ~BpGenAsyncRunner();
     bool Run(int argc, const char* const argv[], volatile bool & running, bool useSignalHandler);
-    uint64_t m_bundleCount;
+    double m_bundleCount;
     uint64_t m_totalBundlesAcked;
 
     OutductFinalStats m_outductFinalStats;
