+++ common/bpcodec/apps/bpgen/src/BpGenAsyncRunner.cpp
@@ -51,7 +51,7 @@ bool BpGenAsyncRunner::Run(int argc, const char* const argv[], volatile bool & r
         m_runningFromSigHandler = true;
         SignalHandler sigHandler(boost::bind(&BpGenAsyncRunner::MonitorExitKeypressThreadFunction, this));
         uint32_t bundleSizeBytes;
-        uint32_t bundleRate;
+        double bundleRate;
         //uint32_t tcpclFragmentSize;
         uint32_t durationSeconds;
         cbhe_eid_t myEid;
@@ -72,7 +72,7 @@ bool BpGenAsyncRunner::Run(int argc, const char* const argv[], volatile bool & r
             desc.add_options()
                 ("help", "Produce help message.")
                 ("bundle-size", boost::program_options::value<uint32_t>()->default_value(100), "Bundle size bytes.")
-                ("bundle-rate", boost::program_options::value<uint32_t>()->default_value(1500), "Bundle rate. (0=>as fast as possible)")
+                ("bundle-rate", boost::program_options::value<double>()->default_value(1500), "Bundle rate. (0=>as fast as possible)")
                 ("duration", boost::program_options::value<uint32_t>()->default_value(0), "Seconds to send bundles for (0=>infinity).")
                 ("my-uri-eid", boost::program_options::value<std::string>()->default_value("ipn:1.1"), "BpGen Source Node Id.")
                 ("dest-uri-eid", boost::program_options::value<std::string>()->default_value("ipn:2.1"), "BpGen sends to this final destination Eid.")
@@ -152,7 +152,7 @@ bool BpGenAsyncRunner::Run(int argc, const char* const argv[], volatile bool & r
             }
 
             bundleSizeBytes = vm["bundle-size"].as<uint32_t>();
-            bundleRate = vm["bundle-rate"].as<uint32_t>();
+            bundleRate = vm["bundle-rate"].as<double>();
             durationSeconds = vm["duration"].as<uint32_t>();
             myCustodianServiceId = vm["my-custodian-service-id"].as<uint64_t>();
             bundleSendTimeoutSeconds = vm["bundle-send-timeout-seconds"].as<unsigned int>();
